requirements_server.txt
uvicorn
fastapi==0.55.1
sqlalchemy-utils
# psycopg2 # for postgresql only
alembic
